+++ Core/CommonUI/Package.swift
@@ -9,21 +9,21 @@ let package = Package(
         .iOS(.v15),
     ],
     products: [
-        // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
             name: "CommonUI",
-            targets: ["CommonUI"]),
+            targets: ["CommonUI"]
+        ),
     ],
     dependencies: [
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
+        .package(path: "../Foundation/Models")
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "CommonUI",
-            dependencies: []),
+            dependencies: [
+                "Models"
+            ]
+        ),
         .testTarget(
             name: "CommonUITests",
             dependencies: ["CommonUI"]),
