+++ Features/Search/Sources/Search/Data/SearchRepository.swift
@@ -1,0 +1,25 @@
+//
+//  SearchRepository.swift
+//  iMovie
+//
+//  Created by Tuan Hoang on 24/06/2023.
+//
+
+import Foundation
+import Network
+import Models
+
+final class SearchRepository: ISearchRepository {
+    private let apiClientService: IAPIClientService
+    init(apiClientService: IAPIClientService) {
+        self.apiClientService = apiClientService
+    }
+
+    func search(keyword: String) async throws -> [Movie] {
+        // TODO (Tuan): To add real search API
+        try await apiClientService.request(
+            APIEndpoints.fetchDiscoveryMovies(),
+            mapper: TrendingMovieResponseMapper()
+        )
+    }
+}
