+++ Foundation/Models/Sources/Models/Genre.swift
@@ -1,0 +1,41 @@
+//
+//  Genre.swift
+//  iMovie
+//
+//  Created by Tuan Hoang on 18/06/2023.
+//
+
+import Foundation
+
+public struct Genre: Decodable, Equatable {
+
+    public enum GenreType: String, Decodable {
+        case action = "Action"
+        case adventure = "Adventure"
+        case aimation = "Animation"
+        case comedy = "Comedy"
+        case crime = "Crime"
+        case documentary = "Documentary"
+        case drama = "Drama"
+        case family = "Family"
+        case fantasy = "Fantasy"
+        case history = "History"
+        case horror = "Horror"
+        case music = "Music"
+        case mystery = "Mystery"
+        case scienceFiction = "Science Fiction"
+        case tvMovie = "TV Movie"
+        case thriller = "Thriller"
+        case war = "War"
+        case western = "Western"
+        case romance = "Romance"
+    }
+
+    public let id: Int
+    public let name: GenreType
+
+    public init(id: Int, name: GenreType) {
+        self.id = id
+        self.name = name
+    }
+}
