+++ README.md
@@ -1,9 +1,5 @@
 #  iMovie
 
 ## Ref:
-- API Design: https://www.figma.com/file/rqjEdJecH0tr6uIyT72KXh/NAP001?type=design&node-id=6-9199&t=Ai7i6CXqzbqM468M-0
+- Figma Design: https://www.figma.com/file/rqjEdJecH0tr6uIyT72KXh/NAP001?type=design&node-id=6-9199&t=Ai7i6CXqzbqM468M-0
 - API Document: https://developer.themoviedb.org/reference/intro/getting-started
-
-## Todo improvement
-- Improve APIEndpoints
-- Create util for Colors
