--- iMovie/Data/DTO/GenresResponse.swift
@@ -1,12 +1,0 @@
-//
-//  GenresResponse.swift
-//  iMovie
-//
-//  Created by Tuan Hoang on 18/06/2023.
-//
-
-import Foundation
-
-struct GenresResponse: Decodable {
-    let genres: [Genre]
-}
